+++ elastiknn-models/src/main/java/com/klibisz/elastiknn/vectors/DefaultFloatVectorOps.java
@@ -3,7 +3,7 @@ package com.klibisz.elastiknn.vectors;
 public final class DefaultFloatVectorOps implements FloatVectorOps {
 
     @Override
-    public float dotProduct(float[] v1, float[] v2) {
+    public double dotProduct(float[] v1, float[] v2) {
         float dp = 0f;
         for (int i = 0; i < v1.length; i++) dp = Math.fma(v1[i], v2[i], dp);
         return dp;
