+++ elastiknn-models/src/main/java/com/klibisz/elastiknn/vectors/FloatVectorOps.java
@@ -2,7 +2,7 @@ package com.klibisz.elastiknn.vectors;
 
 public interface FloatVectorOps {
 
-    float dotProduct(final float[] v1, final float[] v2);
+    double dotProduct(final float[] v1, final float[] v2);
 
     double euclideanDistance(final float[] v1, final float[] v2);
